Move app configs to a folder
